requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
numpy>=1.26.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
pillow>=10.0.0
joblib>=1.2.0
ftfy
opencv-python>=4.8.0
regex>=2023.3.23
scipy>=1.10.0
gunicorn==20.1.0